quotet_api/requirements.txt
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
Django==2.2
django-cors-headers==2.5.2
django-s3-storage==0.12.4
djangorestframework==3.9.2
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
psycopg2==2.8
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
